requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-index==0.9.48
llama-index-vector-stores-chroma==0.1.4
chromadb==0.4.22
openai==1.6.1
python-dotenv==1.0.0
werkzeug==3.0.1
pypdf==3.17.4
python-docx==1.1.0
